+++ RateMonitor/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## v0.3.0
+- Add config UI: Font Size. Default value is 16.
+- Add expandable button (+) in title to quick change rate unit and count multiplier.
+- The count multiplier affect machine count. It resets to 1 after title is folded.
+- When the window is open, Ctrl+right click can continue to add selected buildings (repeatable), and Shift+right click can remove selected buildings.  
+- When pressing hotkey but the selection tool is not available, it will open the mod window with the last selection.
+- The selection tool can now penetrate the mod window to finishing selecting.
+
 ## v0.2.0
 - Support ray receivers and energy exchangers production calculation.  
 - Make rate panel and window resizeable and memorize the settings.